// src/db/nodes/track-layouts/types/InputTrackLayoutCreate.ts
export type InputTrackLayoutCreate = {
    name: string
    year_from: number | null
    year_to: number | null
    length: number | null
    length_unit: string | null
    direction: string | null
    elevation_change: number | null
    elevation_change_unit: string | null
    surface: string | null
}